requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
unstructured